separate 3d cable joint code
